Put some Config stuff into .cc file
